chore: move testhelpers to separate package
