+++ driver/config/testhelpers/config.go
@@ -1,7 +1,7 @@
 // Copyright © 2024 Ory Corp
 // SPDX-License-Identifier: Apache-2.0
 
-package config
+package testhelpers
 
 import (
 	"context"
