+++ internal/testhelpers/config.go
@@ -8,6 +8,8 @@ import (
 	"encoding/base64"
 	"testing"
 
+	confighelpers "github.com/ory/kratos/driver/config/testhelpers"
+
 	"github.com/spf13/pflag"
 	"github.com/stretchr/testify/require"
 
@@ -33,7 +35,7 @@ func DefaultIdentitySchemaConfig(url string) map[string]any {
 }
 
 func WithDefaultIdentitySchema(ctx context.Context, url string) context.Context {
-	return config.WithConfigValues(ctx, DefaultIdentitySchemaConfig(url))
+	return confighelpers.WithConfigValues(ctx, DefaultIdentitySchemaConfig(url))
 }
 
 // Deprecated: Use context-based WithDefaultIdentitySchema instead
@@ -58,7 +60,7 @@ func WithAddIdentitySchema(ctx context.Context, t *testing.T, conf *config.Confi
 	schemas, err := conf.IdentityTraitsSchemas(ctx)
 	require.NoError(t, err)
 
-	return config.WithConfigValue(ctx, config.ViperKeyIdentitySchemas, append(schemas, config.Schema{
+	return confighelpers.WithConfigValue(ctx, config.ViperKeyIdentitySchemas, append(schemas, config.Schema{
 		ID:  id,
 		URL: url,
 	})), id
