+++ selfservice/strategy/code/test/persistence.go
@@ -8,6 +8,8 @@ import (
 	"testing"
 	"time"
 
+	confighelpers "github.com/ory/kratos/driver/config/testhelpers"
+
 	"github.com/ory/kratos/internal/testhelpers"
 	"github.com/ory/kratos/persistence"
 	"github.com/ory/kratos/selfservice/flow"
@@ -31,7 +33,7 @@ func TestPersister(ctx context.Context, p interface {
 	return func(t *testing.T) {
 		nid, p := testhelpers.NewNetworkUnlessExisting(t, ctx, p)
 
-		ctx := config.WithConfigValue(ctx, config.ViperKeySecretsDefault, []string{"secret-a", "secret-b"})
+		ctx := confighelpers.WithConfigValue(ctx, config.ViperKeySecretsDefault, []string{"secret-a", "secret-b"})
 
 		t.Run("code=recovery", func(t *testing.T) {
 			newRecoveryCodeDTO := func(t *testing.T, email string) (*code.CreateRecoveryCodeParams, *recovery.Flow, *identity.RecoveryAddress) {
@@ -50,7 +52,7 @@ func TestPersister(ctx context.Context, p interface {
 				require.NoError(t, p.CreateIdentity(ctx, &i))
 
 				return &code.CreateRecoveryCodeParams{
-					RawCode:         string(randx.MustString(8, randx.Numeric)),
+					RawCode:         randx.MustString(8, randx.Numeric),
 					FlowID:          f.ID,
 					RecoveryAddress: &i.RecoveryAddresses[0],
 					ExpiresIn:       time.Minute,
